+++ editor/editor-wikimarkup-transformer/src/parser/utils/attrs.ts
@@ -1,19 +1,19 @@
-// querystring-es3 is used because the browser implementation of querystring behaves differently than the NodeJS module
-// https://product-fabric.atlassian.net/browse/CS-2801
-import { parse as parseQuery } from 'querystring-es3';
-
 export function parseAttrs(
-  str: string,
+  str?: string,
   sep: string = '|',
 ): { [key: string]: string } {
-  const output = parseQuery(str, sep);
-
-  // take only first value of the same keys
-  Object.keys(output).forEach(key => {
-    if (Array.isArray(output[key])) {
-      output[key] = output[key][0];
+  const output: { [key: string]: string } = {};
+  if (!str) {
+    return output;
+  }
+  const attributesStr = str.split(sep);
+  attributesStr.forEach(attributeStr => {
+    const [key, ...value] = attributeStr.split('=');
+    // take only first value of the same keys
+    if (!output[key]) {
+      output[key] = value.join('=');
     }
   });
 
-  return output as { [key: string]: string };
+  return output;
 }